Broke sigma_sq optimization into numpy and jax implementations. Note that this update is breaking at least one test case.
