+++ MuyGPyS/_src/gp/numpy_muygps.py
@@ -5,6 +5,8 @@
 
 import numpy as np
 
+from typing import Generator
+
 
 def _muygps_compute_solve(
     K: np.ndarray,
@@ -32,3 +34,32 @@ def _muygps_compute_diagonal_variance(
         ).reshape(batch_count, nn_count),
         axis=1,
     )
+
+
+def _muygps_sigma_sq_optim(
+    K: np.ndarray,
+    nn_indices: np.ndarray,
+    targets: np.ndarray,
+    eps: float,
+):
+    batch_count, nn_count = nn_indices.shape
+    _, response_count = targets.shape
+
+    sigma_sq = np.zeros((response_count,))
+    for i in range(response_count):
+        sigma_sq[i] = sum(_get_sigma_sq(K, targets[:, i], nn_indices, eps)) / (
+            nn_count * batch_count
+        )
+    return sigma_sq
+
+
+def _get_sigma_sq(
+    K: np.ndarray,
+    target_col: np.ndarray,
+    nn_indices: np.ndarray,
+    eps: float,
+) -> Generator[float, None, None]:
+    batch_count, nn_count = nn_indices.shape
+    for j in range(batch_count):
+        Y_0 = target_col[nn_indices[j, :]]
+        yield Y_0 @ np.linalg.solve(K[j, :, :] + eps * np.eye(nn_count), Y_0)
